Fix Windows warnings

+++ blazingio.hpp
@@ -153,7 +153,7 @@ struct istream_impl {
         ensure(base)
         ensure(VirtualFree(base, 0, MEM_RELEASE))
         // Map the file there
-        size_t mmaped_region_size = (file_size + 4095) & -65536;
+        DWORD mmaped_region_size = (file_size + 4095) & -65536;
         // If we remove this if and always call CreateFileMapping, it's going to interpret 0 as
         // "max", which we don't want.
         if (mmaped_region_size)
@@ -779,10 +779,10 @@ struct blazingio_ostream {
         DWORD n_written;
         ensure(
             handle == INVALID_HANDLE_VALUE
-                ? WriteFile(stdout_handle, base, (char*)ptr - base, &n_written, NULL)
+                ? WriteFile(stdout_handle, base, DWORD((char*)ptr - base), &n_written, NULL)
                 : (
-                    WriteFile(handle, base, ((char*)ptr - base + 4095) & -4096, &n_written, NULL)
-                    && ~_chsize(1, (char*)ptr - base)
+                    WriteFile(handle, base, DWORD(((char*)ptr - base + 4095) & -4096), &n_written, NULL)
+                    && ~_chsize(1, int((char*)ptr - base))
                 )
         )
 @end
